refactor: Split the signboard into components.
