+++ frontend/src/components/models/employee/EmpTest.tsx
@@ -9,10 +9,9 @@ import {
   calculateVhBasedOnFigma,
   calculateVwBasedOnFigma,
 } from 'utils/calculateSizeBasedOnFigma'
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { faCaretRight } from '@fortawesome/free-solid-svg-icons'
 import { EmployeeOnlineStatusLabel } from 'components/models/employee/EmployeeOnlineStatusLabel'
 import { EmployeeInformation } from 'components/models/employee/EmployeeInformation'
+import { EmployeeSignBoard } from 'components/models/employee/EmployeeSignBoard'
 import { TaskColumnList } from 'components/models/task/TaskColumnList'
 import Drawer from '@mui/material/Drawer'
 import styled from 'styled-components'
@@ -26,7 +25,7 @@ type Props = {
 } & Partial<Groups>
 
 export const EmpTest: React.FC<Props> = ({ groups, lists, onDragEnd, handleAddTask }) => {
-  const [open, setOpen] = React.useState(false)
+  const [isOpen, setIsOpen] = React.useState(false)
 
   return (
     <div style={{ display: 'flex' }}>
@@ -42,7 +41,7 @@ export const EmpTest: React.FC<Props> = ({ groups, lists, onDragEnd, handleAddTa
         }}
         variant="persistent"
         anchor="left"
-        open={open}>
+        open={isOpen}>
         <StyledContaner>
           <StyledMemberContainer>
             <StyledLabelContainer>
@@ -70,12 +69,9 @@ export const EmpTest: React.FC<Props> = ({ groups, lists, onDragEnd, handleAddTa
           </StyledMemberContainer>
         </StyledContaner>
       </Drawer>
-      <StyledMain open={open}>
-        <div style={{ display: 'flex', alignItems: 'baseline' }}>
-          <StyledSignBoardContainer onClick={() => setOpen(!open)}>
-            <StyledH3>MEMBER</StyledH3>
-            <StyledFontAwesomeIcon icon={faCaretRight} />
-          </StyledSignBoardContainer>
+      <StyledMain open={isOpen}>
+        <StyledMainWrap>
+          <EmployeeSignBoard isOpen={isOpen} handleClick={() => setIsOpen(!isOpen)} />
           <DragDropContext onDragEnd={onDragEnd}>
             <Droppable droppableId="board" direction="horizontal" type={DropType.COLUMN}>
               {provided => (
@@ -86,7 +82,7 @@ export const EmpTest: React.FC<Props> = ({ groups, lists, onDragEnd, handleAddTa
               )}
             </Droppable>
           </DragDropContext>
-        </div>
+        </StyledMainWrap>
       </StyledMain>
     </div>
   )
@@ -108,27 +104,10 @@ const StyledMemberContainer = styled.div`
   overflow-x: hidden;
   overflow-y: auto;
 `
-const StyledSignBoardContainer = styled.div`
+const StyledMainWrap = styled.div`
   display: flex;
-  align-items: center;
-  border: 2px solid ${({ theme }) => theme.COLORS.BRANDY};
-  border-radius: 4px;
-  padding: ${calculateMinSizeBasedOnFigmaWidth(4)};
-  background-color: ${({ theme }) => theme.COLORS.MINE_SHAFT};
-  filter: drop-shadow(-4px 4px 2px rgba(0, 0, 0, 0.5));
-  cursor: pointer;
-  writing-mode: vertical-rl;
-  -webkit-writing-mode: vertical-rl;
-  -ms-writing-mode: tb-rl;
-  text-orientation: upright;
-`
-const StyledH3 = styled.h3`
-  font-size: ${({ theme }) => theme.FONT_SIZES.SIZE_18};
-  color: ${({ theme }) => theme.COLORS.WHITE};
-`
-const StyledFontAwesomeIcon = styled(FontAwesomeIcon)`
-  font-size: ${({ theme }) => theme.FONT_SIZES.SIZE_18};
-  color: ${({ theme }) => theme.COLORS.WHITE};
+  align-items: baseline;
+  gap: ${calculateMinSizeBasedOnFigmaWidth(16)};
 `
 const StyledEmployeeContainer = styled.div`
   margin-bottom: ${calculateMinSizeBasedOnFigmaWidth(10)};
